Extract Modal into component
