+++ src/components/FileContributors.js
@@ -1,16 +1,13 @@
-import React, { useState, useRef } from "react"
-import { motion } from "framer-motion"
+import React, { useState } from "react"
 import { useIntl } from "gatsby-plugin-intl"
 import styled from "styled-components"
-import { Twemoji } from "react-emoji-render"
 
 import Button from "./Button"
 import Icon from "./Icon"
 import Link from "./Link"
+import Modal from "./Modal"
 import { FakeButtonSecondary } from "./SharedStyledComponents"
 import { getLocaleTimestamp } from "../utils/time"
-import { useOnClickOutside } from "../hooks/useOnClickOutside"
-import { useKeyPress } from "../hooks/useKeyPress"
 
 const Container = styled.div`
   display: flex;
@@ -98,81 +95,9 @@ const GithubIcon = styled(Icon)`
   margin-right: 0.5rem;
 `
 
-const StyledOverlay = styled(motion.div)`
-  position: fixed;
-  background: rgba(0, 0, 0, 0.7);
-  will-change: opacity;
-  top: 0;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 100%;
-`
-
-// TODO extract Modal into standalone component
-const Overlay = ({ isActive }) => {
-  return (
-    <StyledOverlay
-      initial={false}
-      animate={{ opacity: isActive ? 1 : 0, zIndex: isActive ? 1001 : -1 }}
-      transition={{ duration: 0.2 }}
-    />
-  )
-}
-
-const ModalContainer = styled.div`
-  top: 0px;
-  left: 0px;
-  right: 0px;
-  position: fixed;
-  z-index: 1002;
-  cursor: pointer;
-  padding: 15% 1rem 0;
-  width: 100%;
-  height: 100%;
-`
-
-const Modal = styled.div`
-  padding: 1rem;
-  height: auto;
-  cursor: auto;
-  max-height: 100%;
-  max-width: 600px;
-  background: ${(props) => props.theme.colors.background};
-  display: flex;
-  justify-content: space-between;
-  box-shadow: rgba(0, 0, 0, 0.16) 0px 2px 4px 0px;
-  width: 100%;
-  border-radius: 8px;
-  border: 1px solid rgb(189, 189, 189);
-  margin: 0px auto;
-`
-
-const ModalContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 1rem;
-`
-
-const ModalClose = styled.div`
-  margin: 1rem;
-`
-const ModalCloseIcon = styled(Icon)`
-  cursor: pointer;
-`
-
 const ModalTitle = styled.h2`
   margin-top: 0;
   margin-bottom: 0.5rem;
-`
-
-const Emoji = styled(Twemoji)`
-  & > img {
-    width: 3em !important;
-    height: 3em !important;
-    margin-bottom: 0 !important;
-  }
-  margin-right: 1rem;
 `
 
 const ContributorList = styled.ul`
@@ -191,20 +116,6 @@ const Contributor = styled.li`
 const FileContributors = ({ gitCommits, className, editPath }) => {
   const [isModalOpen, setModalOpen] = useState(false)
   const intl = useIntl()
-  const ref = useRef()
-
-  // Close modal on outside clicks & `Escape` keypress
-  useOnClickOutside(ref, () => setModalOpen(false))
-  useKeyPress(`Escape`, () => setModalOpen(false))
-
-  const toggleModal = (e) => {
-    // If user clicks on "X" icon, close modal
-    if (e.target.tagName === "path") {
-      setModalOpen(false)
-    } else {
-      setModalOpen(true)
-    }
-  }
 
   const commits = gitCommits.map((commit) => {
     return commit.node
@@ -226,35 +137,20 @@ const FileContributors = ({ gitCommits, className, editPath }) => {
 
   return (
     <div className={className}>
-      <Overlay isActive={isModalOpen} />
-      {isModalOpen && (
-        <ModalContainer>
-          <Modal ref={ref}>
-            <ModalContent>
-              <ModalTitle>Contributors</ModalTitle>
-              <div>Everyone who has contributed to this page – thank you!</div>
-              <ContributorList>
-                {uniqueContributors.map((contributor) => {
-                  return (
-                    <Contributor key={contributor.email}>
-                      <Avatar
-                        src={contributor.avatarUrl}
-                        alt={contributor.name}
-                      />
-                      <Link to={contributor.user.url}>
-                        @{contributor.user.login}
-                      </Link>
-                    </Contributor>
-                  )
-                })}
-              </ContributorList>
-            </ModalContent>
-            <ModalClose onClick={toggleModal}>
-              <ModalCloseIcon name="close" />
-            </ModalClose>
-          </Modal>
-        </ModalContainer>
-      )}
+      <Modal isOpen={isModalOpen} setIsOpen={setModalOpen}>
+        <ModalTitle>Contributors</ModalTitle>
+        <div>Everyone who has contributed to this page – thank you!</div>
+        <ContributorList>
+          {uniqueContributors.map((contributor) => {
+            return (
+              <Contributor key={contributor.email}>
+                <Avatar src={contributor.avatarUrl} alt={contributor.name} />
+                <Link to={contributor.user.url}>@{contributor.user.login}</Link>
+              </Contributor>
+            )
+          })}
+        </ContributorList>
+      </Modal>
       <Container>
         <LeftContent>
           <Avatar src={lastContributor.avatarUrl} alt={lastContributor.name} />
@@ -267,7 +163,7 @@ const FileContributors = ({ gitCommits, className, editPath }) => {
           </Info>
         </LeftContent>
         <ButtonContainer>
-          <ContributorsButton onClick={toggleModal}>
+          <ContributorsButton onClick={() => setModalOpen(true)}>
             See contributors
           </ContributorsButton>
           {editPath && (
